// SMS.Web/src/app/api/enhanceEndpoints/tags.ts
export enum Tag {
  Allocation = "Allocation",
  BankAllocation = "BankAllocation",
  AllocationSummary = "AllocationSummary",
  CurrentUser = "CurrentUser",
  ParValue = "ParValue",
  Payment = "Payment",
  Subscription = "Subscription",
  ShareholderInfo = "ShareholderInfo",
  Shareholder = "Shareholder",
  ShareholderContacts = "ShareholderContacts",
  ShareholderAddresses = "ShareholderAddresses",
  SubscriptionGroup = "SubscriptionGroup",
  Transfer = "Transfer",
  Users = "Users",
  UserDetail = "UserDetail",
  Dividend = "Dividend",
}